refactor bundled deps
