+++ modules/packages/ImageHelper/stb/README
@@ -1,0 +1,3 @@
+# README
+
+Bundled stb headers are pulled from https://github.com/nothings/stb
