+++ modules/packages/ImageHelper/stb_image_helper.h
@@ -24,12 +24,12 @@
 #define STBI_ONLY_PNG
 #define STBI_ONLY_JPEG
 #define STBI_NO_STDIO
-#include "stb_image.h"
+#include "stb/stb_image.h"
 
 #define STB_IMAGE_WRITE_IMPLEMENTATION
 #define STBI_WRITE_NO_STDIO
 #define STB_IMAGE_WRITE_STATIC
 
-#include "stb_image_write.h"
+#include "stb/stb_image_write.h"
 
 #endif
